REST API improvements
