+++ test/org/opensolaris/opengrok/web/api/v1/controller/StatsControllerTest.java
@@ -20,7 +20,7 @@
 /*
  * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
  */
-package org.opensolaris.opengrok.web.api.controller;
+package org.opensolaris.opengrok.web.api.v1.controller;
 
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.JerseyTest;
@@ -105,7 +105,7 @@ public class StatsControllerTest extends JerseyTest {
         Response response = target("stats")
                 .path("reload")
                 .request()
-                .post(Entity.text(""));
+                .put(Entity.text(""));
 
         assertEquals(Response.Status.INTERNAL_SERVER_ERROR.getStatusCode(), response.getStatus());
     }
